Remove gear-programs-common
