+++ gear-programs/bridging-payment/Cargo.toml
@@ -4,9 +4,12 @@ version.workspace = true
 edition.workspace = true
 
 [dependencies]
+grc20-gateway = { workspace = true, features = ["std"] }
+
 gstd.workspace = true
 parity-scale-codec.workspace = true
 scale-info.workspace = true
+primitive-types.workspace = true
 
 [build-dependencies]
 gear-wasm-builder.workspace = true
