--- gear-programs/common/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "gear-programs-common"
-version.workspace = true
-edition.workspace = true
-
-[dependencies]
-primitive-types.workspace = true
-gstd = { workspace = true, optional = true }
-
-[features]
-wasm = ["dep:gstd"]
