+++ gear-programs/grc20-gateway/Cargo.toml
@@ -5,6 +5,9 @@ edition.workspace = true
 
 [dependencies]
 gstd.workspace = true
+primitive-types.workspace = true
+parity-scale-codec.workspace = true
+scale-info.workspace = true
 
 [build-dependencies]
 gear-wasm-builder.workspace = true
