+++ gear-programs/grc20-gateway/src/wasm.rs
@@ -1,7 +1,16 @@
-use gstd::prelude::*;
+use super::*;
+use gstd::{msg, prelude::*};
+use primitive_types::U256;
 
 #[no_mangle]
 extern "C" fn init() {}
 
 #[no_mangle]
-extern "C" fn handle() {}
+extern "C" fn handle() {
+    msg::reply(
+        vara2eth::Response {
+            nonce: U256::zero(),
+        },
+        0,
+    );
+}
